# www/swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CompetitiveBackend",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/competitions/{compID}/rewards": {
      "get": {
        "tags": [
          "CompetitionReward"
        ],
        "summary": "Получить информацию о выдаваемых наградах",
        "parameters": [
          {
            "name": "compID",
            "in": "path",
            "description": "Идентификатор соревнования",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешное выполнение",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompetitionRewardDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompetitionRewardDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompetitionRewardDTO"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Соревнование с таким ID не найдено",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Ошибка сервера"
          }
        }
      },
      "post": {
        "tags": [
          "CompetitionReward"
        ],
        "summary": "Добавить выдаваемую награду",
        "parameters": [
          {
            "name": "compID",
            "in": "path",
            "description": "Идентификатор соревнования",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Описание выдаваемой награды",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompetitionRewardDTO"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompetitionRewardDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompetitionRewardDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompetitionRewardDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешное добавление",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CompetitionReward"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetitionReward"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetitionReward"
                }
              }
            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Пользователь не авторизован как администратор",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Ошибка сервера",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/competitions/{compID}/rewards/{rewardID}": {
      "delete": {
        "tags": [
          "CompetitionReward"
        ],
        "summary": "Удалить выдаваемую награду",
        "parameters": [
          {
            "name": "compID",
            "in": "path",
            "description": "Идентификатор соревнования",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "rewardID",
            "in": "path",
            "description": "Идентификатор выдаваемой награды",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Успешное удаление"
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Пользователь не авторизован как администратор",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Награда или соревнование с таким ID не найдены",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Ошибка сервера",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "CompetitionReward"
        ],
        "summary": "Обновить выдаваемую награду",
        "parameters": [
          {
            "name": "compID",
            "in": "path",
            "description": "Идентификатор соревнования",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "rewardID",
            "in": "path",
            "description": "Идентификатор выдаваемой награды",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompetitionRewardDTO"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompetitionRewardDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompetitionRewardDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompetitionRewardDTO"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Успешное удаление"
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Пользователь не авторизован как администратор",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Награда или соревнование с таким ID не найдены",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Ошибка сервера"
          }
        }
      }
    },
...
    "/api/v1/players/me/rewards": {
      "get": {
        "tags": [
          "PlayerRewards"
        ],
        "summary": "Получить информацию о наградах текущего пользователя",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Индекс страницы",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "count",
            "in": "query",
            "description": "Количество наград на одной странице",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешное выполнение",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlayerRewardDTO"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Ошибка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/players/{playerID}/rewards": {
      "get": {
        "tags": [
          "PlayerRewards"
        ],
        "summary": "Получить информацию о наградах игрока",
        "parameters": [
          {
            "name": "playerID",
            "in": "path",
            "description": "Идентификатор игрока",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Индекс страницы",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "count",
            "in": "query",
            "description": "Количество наград на одной странице",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешное выполнение",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlayerRewardDTO"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Пользователь с таким ID не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Ошибка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "PlayerRewards"
        ],
        "summary": "Выдать награду игроку",
        "parameters": [
          {
            "name": "playerID",
            "in": "path",
            "description": "Идентификатор игрока",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "rewardDescriptionID",
            "in": "query",
            "description": "Идентификатор описания награды",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Награда успешно выдана"
          },
          "400": {
            "description": "Неверные параметры запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Игрок или описание награды не найдены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Ошибка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/players/{playerID}/rewards/{rewardID}": {
      "delete": {
        "tags": [
          "PlayerRewards"
        ],
        "summary": "Удалить награду у игрока",
        "parameters": [
          {
            "name": "rewardID",
            "in": "path",
            "description": "Идентификатор награды",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "playerID",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Награда успешно удалена"
          },
          "404": {
            "description": "Награда с указанным ID не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Ошибка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompetitionReward": {
        "type": "object",
        "additionalProperties": false
      },
      "CompetitionRewardDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "rewardDescriptionID": {
            "type": "integer",
            "format": "int32"
          },
          "conditionByRank": {
            "$ref": "#/components/schemas/RankRewardConditionDTO"
          },
          "conditionByPlace": {
            "$ref": "#/components/schemas/PlaceRewardConditionDTO"
          },
          "competitionID": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PlaceRewardConditionDTO": {
        "type": "object",
        "properties": {
          "minPlace": {
            "type": "integer",
            "format": "int32"
          },
          "maxPlace": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "PlayerRewardDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "playerID": {
            "type": "integer",
            "format": "int32"
          },
          "rewardDescriptionID": {
            "type": "integer",
            "format": "int32"
          },
          "grantedCompetitionID": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "grantDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "RankRewardConditionDTO": {
        "type": "object",
        "properties": {
          "minRank": {
            "type": "number",
            "format": "float"
          },
          "maxRank": {
            "type": "number",
            "format": "float"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateCompetitionRewardDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "rewardDescriptionID": {
            "type": "integer",
            "format": "int32"
          },
          "conditionByRank": {
            "$ref": "#/components/schemas/RankRewardConditionDTO"
          },
          "conditionByPlace": {
            "$ref": "#/components/schemas/PlaceRewardConditionDTO"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Put **_ONLY_** your JWT Bearer token on textbox below!",
        "name": "Bearer",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}